refactor: split recipe run cli command
